requirements.txt
argon2-cffi==20.1.0
asgiref==3.3.1
cffi==1.14.4
Django==3.1.4
django-hosts==4.0
Pillow==8.1.0
psycopg2==2.8.6
pycparser==2.20
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
style==1.1.0
update==0.0.1
